# config/swagger.yaml
swagger: '2.0'
info:
  version: 0.0.1
  title: mobile-topup-service-provider
host: 'localhost:8000'
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /order:
    x-swagger-router-controller: Create new topup order
    post:
      description: Create nre topup order
      operationId: createOrder
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/CreateOrderParameters'
      responses:
        '200':
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/CreateOrderResponse'
        default:
          description: Error Occured
          schema:
            $ref: '#/definitions/ErrorResponse'
  /order/checkTrans:
    x-swagger-router-controller: Check order info by transaction id
    get:
      description: Check order info by transaction id
      operationId: checkTrans
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/CheckTransParameters'
      responses:
        '200':
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/checkTransResponse'
        default:
          description: Error Occured
          schema:
            $ref: '#/definitions/ErrorResponse'
parameters:
  CreateOrderParameters:
    name: body
    in: body
    description: The parameters for creating new order
    schema:
      $ref: '#/definitions/OrdersRequest'
  CheckTransParameters:
    in: query
    name: request_id
    description: Transaction id
    schema:
      $ref: '#/definitions/CheckTransRequest'
definitions:
  OrdersRequest:
    required:
      - phone_number
      - amount
      - request_id
    properties:
      phone_number:
        type: string
        example: 0133456789
        description: The phone number
      amount:
        type: integer
        example: 100000
        description: Total payment amount
        enum:
          - 10000
          - 20000
          - 50000
          - 100000
          - 200000
          - 500000
      request_id:
        type: string
        example: a11122ee122eeaads
  CreateOrderResponse:
    required:
      - request_id
      - status
    properties:
      request_id:
        type: string
        example: a123432we12reewer322
        description: The request id
      status:
        type: string
        example: SUCCESS|PENDING|FAIL
  CheckTransRequest:
    required:
      - request_id
    properties:
      request_id:
        type: string
        example: asasfs123132
  checkTransResponse:
    required:
      - request_id
      - status
    properties:
      request_id:
        type: string
        example: a123432we12reewer322
      status:
        type: string
        example: SUCCESS|PENDING|FAIL
        description: The code which reflect to request status
  ErrorResponse:
    required:
      - status
    properties:
      status:
        type: string
        example: SUCCESS|PENDING|FAIL
      message:
        type: string